refactor(gui): integrate original download.py verbatim; remove duplicate builders
