+++ streamsaavy_app/ui.py
@@ -24,8 +24,8 @@ from tkinter import (
 from tkinter import ttk
 from typing import Dict, Optional
 
-from .download_engine import run_download
 from .downloader import DownloadMode
+from download import run_download
 
 
 class StreamSaavyApp(Tk):
